feat(notifier): edit provider
